# examples/go-server-server-generated/go/model_juegos_response.go
/*
 * Estándar Para Las APIs De Exportación De Datos De Lotería
 *
 * Especificación para las APIs de exportación de datos de lotería, diseñada para ser implementada por sistemas heterogéneos, soportando formatos JSON y XML.   El objetivo es homogeneizar la forma en que los datos son expuestos.  ## Características principales: - Soporte dual para JSON y XML - Autenticación Bearer Token - Filtrado flexible por entidad, moneda, fechas y tipo de lotería - Validación robusta de parámetros  ## Límites importantes: - Solo se pueden consultar datos hasta el día anterior (fecha actual -1) - fechaDesde debe ser menor o igual a fechaHasta - Token de acceso válido por 7 días (604800 segundos) 
 *
 * API version: 1.2
 * Contact: ben.gonzalez@example.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type JuegosResponse struct {
	// Lista de todos los juegos disponibles
	Juegos []string `json:"juegos,omitempty"`
}