# examples/jaxrs-cxf-server-generated/src/gen/java/io/swagger/model/ErrorResponse.java
package io.swagger.model;

import io.swagger.model.ErrorResponseError;
import javax.validation.constraints.*;

import io.swagger.v3.oas.annotations.media.Schema;
import javax.xml.bind.annotation.XmlElement;
import javax.xml.bind.annotation.XmlRootElement;
import javax.xml.bind.annotation.XmlAccessType;
import javax.xml.bind.annotation.XmlAccessorType;
import javax.xml.bind.annotation.XmlType;
import javax.xml.bind.annotation.XmlEnum;
import javax.xml.bind.annotation.XmlEnumValue;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonValue;
import com.fasterxml.jackson.annotation.JsonCreator;

public class ErrorResponse   {
  
  @Schema(description = "")
  private ErrorResponseError error = null;
 /**
   * Get error
   * @return error
  **/
  @JsonProperty("error")
  @NotNull
  public ErrorResponseError getError() {
    return error;
  }

  public void setError(ErrorResponseError error) {
    this.error = error;
  }

  public ErrorResponse error(ErrorResponseError error) {
    this.error = error;
    return this;
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class ErrorResponse {\n");
    
    sb.append("    error: ").append(toIndentedString(error)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private static String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}